chore: move getPermitUtilsInstance to its own file
